typeset example styling for new grid
