+++ packages/addons-website/src/components/TypesetExample/TypesetExample-story.js
@@ -2,6 +2,560 @@ import { storiesOf } from '@storybook/react';
 import React from 'react';
 import TypesetExample from './TypesetExample'
 
+// This is basically the $ibm-type-scale in
+// @ibmduo/type/scss/_variables.scss
+// converted to a js object.
+//
+// Structure kept as close as possible in order to make
+// updating the values easy.
+const ibmTypeScale = {
+  'caption-01': {
+    sm: {
+      'step': 1,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': .75,
+      'line-height': 1,
+      'letter-spacing': 0.32,
+    }
+  },
+  'label-01': {
+    sm: {
+      'step': 1,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': .75,
+      'line-height': 1,
+      'letter-spacing': 0.32,
+    }
+  },
+  'helper-text-01': {
+    sm: {
+      'step': 1,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-style': 'italic',
+      'font-size': .75,
+      'line-height': 1,
+      'letter-spacing': 0.32,
+    }
+  },
+
+  'body-short-01': {
+    sm: {
+      'step': 2,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': .875,
+      'line-height': 1.125,
+      'letter-spacing': 0.16,
+    }
+  },
+  'body-short-02': {
+    sm: {
+      'step': 3,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 1,
+      'line-height': 1.375,
+      'letter-spacing': 0,
+    }
+  },
+  'body-long-01': {
+    sm: {
+      'step': 2,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 0.875,
+      'line-height': 1.25,
+      'letter-spacing': 0.16,
+    }
+  },
+  'body-long-02': {
+    sm: {
+      'step': 3,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 1,
+      'line-height': 1.5,
+      'letter-spacing': 0,
+    }
+  },
+  'code-01': {
+    sm: {
+      'step': 1,
+      'font': 'PlexMono',
+      'font-weight': '400',
+      'font-size': .75,
+      'line-height': 1,
+      'letter-spacing': 0.32,
+    }
+  },
+  'code-02': {
+    sm: {
+      'step': 1,
+      'font': 'PlexMono',
+      'font-weight': '400',
+      'font-size': .875,
+      'line-height': 1.25,
+      'letter-spacing': 0.32,
+    }
+  },
+  'heading-01': {
+    sm: {
+      'step': 2,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 0.875,
+      'line-height': 1.125,
+      'letter-spacing': 0.16,
+    },
+  },
+  'heading-02': {
+    sm: {
+      'step': 3,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 1,
+      'line-height': 1.375,
+      'letter-spacing': 0,
+    }
+  },
+  'heading-03': {
+    sm: {
+      'step': 3,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 1.25,
+      'line-height': 1.625,
+      'letter-spacing': '0',
+    }
+  },
+  'productive-heading-04': {
+    sm: {
+      'step': 7,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 1.75,
+      'line-height': 2.25,
+      'letter-spacing': '0',
+    }
+  },
+  'productive-heading-05': {
+    sm: {
+      'step': 8,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 2,
+      'line-height': 2.5,
+      'letter-spacing': '0',
+    }
+  },
+
+  'expressive-heading-04': {
+    sm: {
+      'step': 7,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 1.75,
+      'line-height': 2.25,
+      'letter-spacing': '0',
+    },
+    md: {
+      'step': 7,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 1.75,
+      'line-height': 2.25,
+      'letter-spacing': '0',
+    },
+    lg: {
+      'step': 7,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 1.75,
+      'line-height': 2.25,
+      'letter-spacing': '0',
+    },
+    xlg: {
+      'step': 8,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 2,
+      'line-height': 2.5,
+      'letter-spacing': '0',
+    },
+    max: {
+      'step': 8,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 2,
+      'line-height': 2.5,
+      'letter-spacing': '0',
+    },
+    
+
+  },
+
+  'expressive-heading-05': {
+    sm: {
+      'step': 8,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 2,
+      'line-height': 2.5,
+      'letter-spacing': '0',
+    },
+    md: {
+      'step': 9,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 2.25,
+      'line-height': 2.75,
+      'letter-spacing': '0',
+    },
+    lg: {
+      'step': 10,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 2.625,
+      'line-height': 3.125,
+      'letter-spacing': '0',
+    },
+    xlg: {
+      'step': 11,
+      'font': 'PlexSans',
+      'font-weight': '400',
+      'font-size': 3,
+      'line-height': 3.5,
+      'letter-spacing': '0',
+    },
+    max: {
+      'step': 13,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 3.75,
+      'line-height': 4.375,
+      'letter-spacing': '0',
+    },
+  },
+  'expressive-paragraph-01': {
+    sm: {
+      'step': 6,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 1.5,
+      'line-height': 1.875,
+      'letter-spacing': '0',
+    },
+    md: {
+      'step': 6,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 1.5,
+      'line-height': 1.875,
+      'letter-spacing': '0',
+    },
+    lg: {
+      'step': 7,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 1.75,
+      'line-height': 2.25,
+      'letter-spacing': '0',
+    },
+    xlg: {
+      'step': 7,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 1.75,
+      'line-height': 2.25,
+      'letter-spacing': '0',
+    },
+    max: {
+      'step': 8,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 2,
+      'line-height': 2.5,
+      'letter-spacing': '0',
+    },
+  },
+  'quotation-01': {
+    sm: {
+      'step': 5,
+      'font': 'PlexSerif',
+      'font-weight': '400',
+      'font-size': 1.25,
+      'line-height': 1.625,
+      'letter-spacing': '0',
+    },
+    md: {
+      'step': 5,
+      'font': 'PlexSerif',
+      'font-weight': '400',
+      'font-size': 1.25,
+      'line-height': 1.625,
+      'letter-spacing': '0',
+    },
+    lg: {
+      'step': 6,
+      'font': 'PlexSerif',
+      'font-weight': '400',
+      'font-size': 1.5,
+      'line-height': 1.875,
+      'letter-spacing': '0',
+    },
+    xlg: {
+      'step': 7,
+      'font': 'PlexSerif',
+      'font-weight': '300',
+      'font-size': 1.75,
+      'line-height': 2.25,
+      'letter-spacing': '0',
+    },
+    max: {
+      'step': 8,
+      'font': 'PlexSerif',
+      'font-weight': '300',
+      'font-size': 2,
+      'line-height': 2.5,
+      'letter-spacing': '0',
+    },
+  },
+  'quotation-02': {
+    sm: {
+      'step': 8,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 2,
+      'line-height': 2.5,
+      'letter-spacing': '0',
+    },
+    md: {
+      'step': 9,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 2.25,
+      'line-height': 2.75,
+      'letter-spacing': '0',
+    },
+    lg: {
+      'step': 10,
+      'font': 'PlexSerif',
+      'font-weight': '300',
+      'font-size': 2.625,
+      'line-height': 3.125,
+      'letter-spacing': '0',
+    },
+    xlg: {
+      'step': 11,
+      'font': 'PlexSerif',
+      'font-weight': '300',
+      'font-size': 3,
+      'line-height': 3.5,
+      'letter-spacing': '0',
+    },
+    max: {
+      'step': 13,
+      'font': 'PlexSerif',
+      'font-weight': '300',
+      'font-size': 3.75,
+      'line-height': 4.375,
+      'letter-spacing': '0',
+    },
+  },
+  'display-01': {
+    sm: {
+      'step': 10,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 2.625,
+      'line-height': 3.125,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    md: {
+      'step': 10,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 2.625,
+      'line-height': 3.125,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    lg: {
+      'step': 12,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 3.375,
+      'line-height': 4,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    xlg: {
+      'step': 13,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 3.75,
+      'line-height': 4.375,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    max: {
+      'step': 15,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 4.75,
+      'line-height': 5.375,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+  },
+  'display-02': {
+    sm: {
+      'step': 10,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 2.625,
+      'line-height': 3.125,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    md: {
+      'step': 10,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 2.625,
+      'line-height': 3.125,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    lg: {
+      'step': 12,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 3.375,
+      'line-height': 4,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    xlg: {
+      'step': 13,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 3.75,
+      'line-height': 4.375,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    max: {
+      'step': 15,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 4.75,
+      'line-height': 5.375,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+  },
+  'display-03': {
+    sm: {
+      'step': 10,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 2.625,
+      'line-height': 3.125,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    md: {
+      'step': 14,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 4.25,
+      'line-height': 4.875,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    lg: {
+      'step': 17,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 5.75,
+      'line-height': 6.375,
+      'letter-spacing': '-0.64',
+      'warning': 'Never use this style as the main headline'
+    },
+    xlg: {
+      'step': 20,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 7.625,
+      'line-height': 8.125,
+      'letter-spacing': '-0.64',
+      'warning': 'Never use this style as the main headline'
+    },
+    max: {
+      'step': 23,
+      'font': 'PlexSans',
+      'font-weight': '300',
+      'font-size': 9.75,
+      'line-height': 10.25,
+      'letter-spacing': '-0.96',
+      'warning': 'Never use this style as the main headline'
+    },
+  },
+  'display-04': {
+    sm: {
+      'step': 10,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 2.625,
+      'line-height': 3.125,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    md: {
+      'step': 14,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 4.25,
+      'line-height': 4.875,
+      'letter-spacing': '0',
+      'warning': 'Never use this style as the main headline'
+    },
+    lg: {
+      'step': 17,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 5.75,
+      'line-height': 6.375,
+      'letter-spacing': -0.64,
+      'warning': 'Never use this style as the main headline'
+    },
+    xlg: {
+      'step': 20,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 7.625,
+      'line-height': 8.125,
+      'letter-spacing': -0.64,
+      'warning': 'Never use this style as the main headline'
+    },
+    max: {
+      'step': 23,
+      'font': 'PlexSans',
+      'font-weight': '600',
+      'font-size': 9.75,
+      'line-height': 10.25,
+      'letter-spacing': -0.96,
+      'warning': 'Never use this style as the main headline'
+    },
+  },
+
+}
+
 const typeSets = {
   caption: [
     {
@@ -152,7 +706,8 @@ storiesOf('Typeset Example', module).add(
     <div style={{width: '100%'}}>
       <TypesetExample simulatedScreenWidth={400}
               name="caption"
-              typeSet={typeSets['caption']} >
+              typeSet={typeSets['fluidHeading']}
+              typeScale={ibmTypeScale['expressive-heading-04']} >
 
       </TypesetExample>
     </div>
