+++ packages/addons-website/src/components/TypesetExample/TypesetExample.js
@@ -1,592 +1,24 @@
 import React from 'react'
 import { settings } from 'carbon-components';
-const { prefix } = settings;
 import classnames from 'classnames'
-
 import { findKey, values } from 'lodash'
+import { baseFontSize, breakpoints as carbonBreakpoints } from "@carbon/layout";
 
-const TYPE_VERSION_REPLACE_REGEX = /.*--/
+const { prefix } = settings;
 
-const rem = 16
-
-// responsive
 const breakpoints = {
-  sm: 20 * 16,
-  md: 42 * 16,
-  lg: 66 * 16,
-  xlg: 82 * 16, // No idea why this is stored separately in gridish
-  max: 99 * 16,
-}
-
-// This is not part of ibmduo-grid, only used in this website
-const breakpointsREM = {
-  sm: 20 + 'rem',
-  md: 42 + 'rem',
-  lg: 66 + 'rem',
-  xlg: 82 + 'rem',
-  max: 99 + 'rem',
+  sm: Number(carbonBreakpoints.sm.width.replace('rem', '')) * baseFontSize,
+  md: Number(carbonBreakpoints.md.width.replace('rem', '')) * baseFontSize,
+  lg: Number(carbonBreakpoints.lg.width.replace('rem', '')) * baseFontSize,
+  xlg: Number(carbonBreakpoints.xlg.width.replace('rem', '')) * baseFontSize,
+  max: Number(carbonBreakpoints.max.width.replace('rem', '')) * baseFontSize,
 }
 
 const defaultTypeValues = {
   'letter-spacing': 0,
 }
 
-// This is basically the $ibm-type-scale in
-// @ibmduo/type/scss/_variables.scss
-// converted to a js object.
-//
-// Structure kept as close as possible in order to make
-// updating the values easy.
-const ibmTypeScale = {
-  'caption-01': {
-    sm: {
-      'step': 1,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': .75,
-      'line-height': 1,
-      'letter-spacing': 0.32,
-    }
-  },
-  'label-01': {
-    sm: {
-      'step': 1,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': .75,
-      'line-height': 1,
-      'letter-spacing': 0.32,
-    }
-  },
-  'helper-text-01': {
-    sm: {
-      'step': 1,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-style': 'italic',
-      'font-size': .75,
-      'line-height': 1,
-      'letter-spacing': 0.32,
-    }
-  },
-
-  'body-short-01': {
-    sm: {
-      'step': 2,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': .875,
-      'line-height': 1.125,
-      'letter-spacing': 0.16,
-    }
-  },
-  'body-short-02': {
-    sm: {
-      'step': 3,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 1,
-      'line-height': 1.375,
-      'letter-spacing': 0,
-    }
-  },
-  'body-long-01': {
-    sm: {
-      'step': 2,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 0.875,
-      'line-height': 1.25,
-      'letter-spacing': 0.16,
-    }
-  },
-  'body-long-02': {
-    sm: {
-      'step': 3,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 1,
-      'line-height': 1.5,
-      'letter-spacing': 0,
-    }
-  },
-  'code-01': {
-    sm: {
-      'step': 1,
-      'font': 'PlexMono',
-      'font-weight': '400',
-      'font-size': .75,
-      'line-height': 1,
-      'letter-spacing': 0.32,
-    }
-  },
-  'code-02': {
-    sm: {
-      'step': 1,
-      'font': 'PlexMono',
-      'font-weight': '400',
-      'font-size': .875,
-      'line-height': 1.25,
-      'letter-spacing': 0.32,
-    }
-  },
-  'heading-01': {
-    sm: {
-      'step': 2,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 0.875,
-      'line-height': 1.125,
-      'letter-spacing': 0.16,
-    },
-  },
-  'heading-02': {
-    sm: {
-      'step': 3,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 1,
-      'line-height': 1.375,
-      'letter-spacing': 0,
-    }
-  },
-  'heading-03': {
-    sm: {
-      'step': 3,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 1.25,
-      'line-height': 1.625,
-      'letter-spacing': '0',
-    }
-  },
-  'productive-heading-04': {
-    sm: {
-      'step': 7,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 1.75,
-      'line-height': 2.25,
-      'letter-spacing': '0',
-    }
-  },
-  'productive-heading-05': {
-    sm: {
-      'step': 8,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 2,
-      'line-height': 2.5,
-      'letter-spacing': '0',
-    }
-  },
-
-  'expressive-heading-04': {
-    sm: {
-      'step': 7,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 1.75,
-      'line-height': 2.25,
-      'letter-spacing': '0',
-    },
-    md: {
-      'step': 7,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 1.75,
-      'line-height': 2.25,
-      'letter-spacing': '0',
-    },
-    lg: {
-      'step': 7,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 1.75,
-      'line-height': 2.25,
-      'letter-spacing': '0',
-    },
-    xlg: {
-      'step': 8,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 2,
-      'line-height': 2.5,
-      'letter-spacing': '0',
-    },
-    max: {
-      'step': 8,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 2,
-      'line-height': 2.5,
-      'letter-spacing': '0',
-    },
-    
-
-  },
-
-  'expressive-heading-05': {
-    sm: {
-      'step': 8,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 2,
-      'line-height': 2.5,
-      'letter-spacing': '0',
-    },
-    md: {
-      'step': 9,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 2.25,
-      'line-height': 2.75,
-      'letter-spacing': '0',
-    },
-    lg: {
-      'step': 10,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 2.625,
-      'line-height': 3.125,
-      'letter-spacing': '0',
-    },
-    xlg: {
-      'step': 11,
-      'font': 'PlexSans',
-      'font-weight': '400',
-      'font-size': 3,
-      'line-height': 3.5,
-      'letter-spacing': '0',
-    },
-    max: {
-      'step': 13,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 3.75,
-      'line-height': 4.375,
-      'letter-spacing': '0',
-    },
-  },
-  'expressive-paragraph-01': {
-    sm: {
-      'step': 6,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 1.5,
-      'line-height': 1.875,
-      'letter-spacing': '0',
-    },
-    md: {
-      'step': 6,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 1.5,
-      'line-height': 1.875,
-      'letter-spacing': '0',
-    },
-    lg: {
-      'step': 7,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 1.75,
-      'line-height': 2.25,
-      'letter-spacing': '0',
-    },
-    xlg: {
-      'step': 7,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 1.75,
-      'line-height': 2.25,
-      'letter-spacing': '0',
-    },
-    max: {
-      'step': 8,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 2,
-      'line-height': 2.5,
-      'letter-spacing': '0',
-    },
-  },
-  'quotation-01': {
-    sm: {
-      'step': 5,
-      'font': 'PlexSerif',
-      'font-weight': '400',
-      'font-size': 1.25,
-      'line-height': 1.625,
-      'letter-spacing': '0',
-    },
-    md: {
-      'step': 5,
-      'font': 'PlexSerif',
-      'font-weight': '400',
-      'font-size': 1.25,
-      'line-height': 1.625,
-      'letter-spacing': '0',
-    },
-    lg: {
-      'step': 6,
-      'font': 'PlexSerif',
-      'font-weight': '400',
-      'font-size': 1.5,
-      'line-height': 1.875,
-      'letter-spacing': '0',
-    },
-    xlg: {
-      'step': 7,
-      'font': 'PlexSerif',
-      'font-weight': '300',
-      'font-size': 1.75,
-      'line-height': 2.25,
-      'letter-spacing': '0',
-    },
-    max: {
-      'step': 8,
-      'font': 'PlexSerif',
-      'font-weight': '300',
-      'font-size': 2,
-      'line-height': 2.5,
-      'letter-spacing': '0',
-    },
-  },
-  'quotation-02': {
-    sm: {
-      'step': 8,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 2,
-      'line-height': 2.5,
-      'letter-spacing': '0',
-    },
-    md: {
-      'step': 9,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 2.25,
-      'line-height': 2.75,
-      'letter-spacing': '0',
-    },
-    lg: {
-      'step': 10,
-      'font': 'PlexSerif',
-      'font-weight': '300',
-      'font-size': 2.625,
-      'line-height': 3.125,
-      'letter-spacing': '0',
-    },
-    xlg: {
-      'step': 11,
-      'font': 'PlexSerif',
-      'font-weight': '300',
-      'font-size': 3,
-      'line-height': 3.5,
-      'letter-spacing': '0',
-    },
-    max: {
-      'step': 13,
-      'font': 'PlexSerif',
-      'font-weight': '300',
-      'font-size': 3.75,
-      'line-height': 4.375,
-      'letter-spacing': '0',
-    },
-  },
-  'display-01': {
-    sm: {
-      'step': 10,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 2.625,
-      'line-height': 3.125,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    md: {
-      'step': 10,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 2.625,
-      'line-height': 3.125,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    lg: {
-      'step': 12,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 3.375,
-      'line-height': 4,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    xlg: {
-      'step': 13,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 3.75,
-      'line-height': 4.375,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    max: {
-      'step': 15,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 4.75,
-      'line-height': 5.375,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-  },
-  'display-02': {
-    sm: {
-      'step': 10,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 2.625,
-      'line-height': 3.125,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    md: {
-      'step': 10,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 2.625,
-      'line-height': 3.125,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    lg: {
-      'step': 12,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 3.375,
-      'line-height': 4,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    xlg: {
-      'step': 13,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 3.75,
-      'line-height': 4.375,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    max: {
-      'step': 15,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 4.75,
-      'line-height': 5.375,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-  },
-  'display-03': {
-    sm: {
-      'step': 10,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 2.625,
-      'line-height': 3.125,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    md: {
-      'step': 14,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 4.25,
-      'line-height': 4.875,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    lg: {
-      'step': 17,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 5.75,
-      'line-height': 6.375,
-      'letter-spacing': '-0.64',
-      'warning': 'Never use this style as the main headline'
-    },
-    xlg: {
-      'step': 20,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 7.625,
-      'line-height': 8.125,
-      'letter-spacing': '-0.64',
-      'warning': 'Never use this style as the main headline'
-    },
-    max: {
-      'step': 23,
-      'font': 'PlexSans',
-      'font-weight': '300',
-      'font-size': 9.75,
-      'line-height': 10.25,
-      'letter-spacing': '-0.96',
-      'warning': 'Never use this style as the main headline'
-    },
-  },
-  'display-04': {
-    sm: {
-      'step': 10,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 2.625,
-      'line-height': 3.125,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    md: {
-      'step': 14,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 4.25,
-      'line-height': 4.875,
-      'letter-spacing': '0',
-      'warning': 'Never use this style as the main headline'
-    },
-    lg: {
-      'step': 17,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 5.75,
-      'line-height': 6.375,
-      'letter-spacing': -0.64,
-      'warning': 'Never use this style as the main headline'
-    },
-    xlg: {
-      'step': 20,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 7.625,
-      'line-height': 8.125,
-      'letter-spacing': -0.64,
-      'warning': 'Never use this style as the main headline'
-    },
-    max: {
-      'step': 23,
-      'font': 'PlexSans',
-      'font-weight': '600',
-      'font-size': 9.75,
-      'line-height': 10.25,
-      'letter-spacing': -0.96,
-      'warning': 'Never use this style as the main headline'
-    },
-  },
-
-}
-
 const TypesetExample = props => {
-   
   return (
     <div>
       {(props.typeSet || []).map(type => {
@@ -623,10 +55,10 @@ const TypesetExample = props => {
         //
         // We still need the structured information for UI display purposes though.
         const getCurrentCompoundStylesForBreakpoint = breakpointName => {
-          const typeKeys = Object.keys(breakpointsREM)
+          const typeKeys = Object.keys(breakpoints)
           const typeStylesUntilCurrentBreakpoint = []
           for (let item of typeKeys) {
-            typeStylesUntilCurrentBreakpoint.push(ibmTypeScale[type.key][item])
+            typeStylesUntilCurrentBreakpoint.push(props.typeScale[item])
             if (item === breakpointName) break
           }
           return Object.assign(
@@ -649,16 +81,13 @@ const TypesetExample = props => {
           (nextLargerBreakpointPx - currentBreakpointPx)
 
           const calculateFluidTypeSize = attribute => {
-            // console.log('currentBreakpointSpecs[attribute] ' + currentBreakpointSpecs[attribute])
-            // console.log('nextLargerBreakpointSpecs[attribute] ' + nextLargerBreakpointSpecs[attribute])
-            // console.log('fluidTypeGrowthFactor ' + fluidTypeGrowthFactor)
             return (
-              currentBreakpointSpecs[attribute] * (rem)
+              currentBreakpointSpecs[attribute] * (baseFontSize)
             )
           }
           const calculateFluidLineHeight = attribute => {
             return (
-              currentBreakpointSpecs[attribute] * (rem)
+              currentBreakpointSpecs[attribute] * (baseFontSize)
             )
           }
   
@@ -703,35 +132,28 @@ const TypesetExample = props => {
           warning: currentBreakpointSpecs['warning'],
         }
 
-        const versionClassName = type.version ? `ibm-type-${type.version}` : ''
+        const versionClassName = type.version ? `${prefix}--type-${type.version}` : ''
 
-        // const samplePStyle = css`
-        //   @media (min-width: ${breakpointsREM.md}) {
-        //     padding-right: 25%;
-        //   }
-        // `
-        const versionClassNames = classnames(`ibm-type-${type.key}`, versionClassName) //, samplePStyle)
+        const versionClassNames = classnames(`${prefix}--type-${type.key}`, versionClassName) //, samplePStyle)
 
         return (
           <div
             key={`${props.name}${type.key}${type.version}`}
             style={{ padding: 0 }}
-            className={`${prefix}--typeset-example-container ibm-grid`}
+            className={`${prefix}--typeset-example-container`}
           >
-            <div className="ibm-grid">
-              <div className={`${prefix}--typeset-example-mask ibm-padding`}>
-                <div className={`${prefix}--typeset-example-example`}>
-                  <p
-                    className={versionClassNames}
-                    style={specs}
-                  >
-                    {type.description}
-                  </p>
-                </div>
+            <div className={`${prefix}--typeset-example-row ${prefix}--row`}>
+              <div className={`${prefix}--typeset-example-description ${prefix}--col-md-8`}>
+                <p
+                  className={versionClassNames}
+                  style={specs}
+                >
+                  {type.description}
+                </p>
               </div>
-              <div className={`${prefix}--typeset-example-specs ibm-padding`}>
+              <div className={`${prefix}--typeset-example-specs ${prefix}--col-md-4 ibm-padding`}>
                 <p className={`${prefix}--type-body-short-01`}>
-                  <span className="ibm-type-semibold">
+                  <span className={`${prefix}--type-semibold`}>
                     {type.name}{' '}
                     {/* {type.version &&
                       type.version.replace(TYPE_VERSION_REPLACE_REGEX, '')} */}
@@ -750,11 +172,11 @@ const TypesetExample = props => {
                   <br />
                   Letter-spacing: {displaySpecs['letterSpacing']}rem
                   {displaySpecs['warning'] != null ? (
-                    <span><br /><span className="ibm-type-semibold">warning: </span>{ displaySpecs['warning'] }<br /></span>
+                    <span><br /><span className={`${prefix}--type-semibold`}>warning: </span>{ displaySpecs['warning'] }<br /></span>
                   ) : (
                     <br />
                   )}
-                  <code className={`${prefix}--typeset-example-code-style`}>${type.name}</code>
+                  <code className={`${prefix}--snippet--inline`}>${type.name}</code>
                 </p>
               </div>
             </div>
